Rules:  The chessboard is a 8x8 standard  board. The kings can move from one cell to another if they have an edge or corner in common. If the generated coordinates of the second king are exactly equal to the first, then the second king wins. Both players play optimally. The game is considered to be won by the first king piece if after it moves, then it can have the same coordinates as the second. The second king piece becomes the winner if after this piece is moved, then both kings have the same coordinates. The game is drawn if there is no possible ways for anyone to win if both the pieces are played optimally.
Rules:  The chessboard is a 8x8 standard  board. The kings can move from one cell to another if they have an edge or corner in common. If the generated coordinates of the second king are exactly equal to the first, then the second king wins. Both players play optimally. The game is considered to be won by the first king piece if after it moves, then it can have the same coordinates as the second. The second king piece becomes the winner if after this piece is moved, then both kings have the same coordinates. The game is drawn if there is no possible ways for anyone to win if both the pieces are played optimally.